Updates the About Page
This commit adds a subheading under the total record count to show the
total number of records with errors. The table for harvests now has two
more columns. One for the harvest type and the other is the error count
of the records.

This commit also adds packages to support a linter for the project.